+++ src/main/java/com/example/mentoringproject/post/post/entity/Category.java
@@ -1,4 +1,4 @@
-package com.example.mentoringproject.post.entity;
+package com.example.mentoringproject.post.post.entity;
 
 public enum Category {
   홍보해요, 소통해요, 후기남겨요
