+++ src/main/java/com/example/mentoringproject/post/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.example.mentoringproject.post.repository;
+package com.example.mentoringproject.post.post.repository;
 
-import com.example.mentoringproject.post.entity.Post;
+import com.example.mentoringproject.post.post.entity.Post;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
